simplify config literal parsing and split doctor tests
